Resource takes page extends DragTakeFragment
