+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resourcetakes/view/RecordableTab.kt
@@ -4,6 +4,7 @@ import javafx.scene.control.Tab
 import javafx.scene.control.TabPane
 import org.wycliffeassociates.otter.common.domain.content.Recordable
 import org.wycliffeassociates.otter.jvm.app.ui.resourcetakes.viewmodel.TabRecordableViewModel
+import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
 import tornadofx.*
 import kotlin.math.min
 
@@ -18,7 +19,7 @@ class RecordableTab(
     init {
         textProperty().bind(viewModel.labelProperty)
 
-        RecordableTabContent(viewModel).apply {
+        RecordableTabContentFragment(viewModel).apply {
             formattedTextProperty.bind(viewModel.getFormattedTextBinding())
             this@RecordableTab.content = this.root
         }
@@ -29,7 +30,7 @@ class RecordableTab(
             }
         }
 
-        viewModel.recordableProperty.onChange { item ->
+        viewModel.recordableProperty.onChangeAndDoNow { item ->
             item?.let {
                 checkAndAddSelf()
             } ?: removeSelf()
